Refactoring zkfold-cardano
