+++ cabal.project
@@ -10,13 +10,13 @@ repository cardano-haskell-packages
     d4a35cd3121aa00d18544bb0ac01c3e1691d618f462c46129271bccf39f7e8ee
 
 index-state:
-  , hackage.haskell.org 2024-06-01T13:39:47Z
-  , cardano-haskell-packages 2024-05-29T10:15:00Z
+  , hackage.haskell.org 2024-07-01T10:41:48Z
+  , cardano-haskell-packages 2024-06-28T18:16:58Z
 
 source-repository-package
   type: git
   location: https://github.com/zkFold/zkfold-base.git
-  tag: 170755b18e7c81f9ee777f27976e1c8654fcfe3a
+  tag: 83c06752efa860361eee705a6ea93485272c6315
 
 packages: .
