+++ src/ZkFold/Cardano/Plonk.hs
@@ -1,17 +1,19 @@
-{-# LANGUAGE TypeFamilies   #-}
+{-# LANGUAGE TypeFamilies #-}
 {-# OPTIONS_GHC -Wno-incomplete-uni-patterns #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module ZkFold.Cardano.Plonk where
 
-import           GHC.ByteOrder                            (ByteOrder (..))
+import           GHC.ByteOrder                             (ByteOrder (..))
 import           PlutusTx.Builtins
-import           PlutusTx.Prelude                         (Bool (..), ($), (.), (<>))
-import           Prelude                                  (undefined)
+import           PlutusTx.Prelude                          (Bool (..), ($), (.), (<>), (&&), Eq (..))
+import           Prelude                                   (undefined)
 
 import           ZkFold.Base.Algebra.Basic.Class
-import           ZkFold.Base.Protocol.NonInteractiveProof (NonInteractiveProof (..))
-import           ZkFold.Cardano.Plonk.OnChain
+import           ZkFold.Base.Protocol.NonInteractiveProof  (NonInteractiveProof (..))
+import           ZkFold.Cardano.Plonk.OnChain.BLS12_381.F  (F (..), powTwo)
+import           ZkFold.Cardano.Plonk.OnChain.BLS12_381.G1 (mul)
+import           ZkFold.Cardano.Plonk.OnChain.Data         (InputBytes (..), ProofBytes (..), SetupBytes (..))
 
 data PlonkPlutus
 
@@ -34,9 +36,8 @@ instance NonInteractiveProof PlonkPlutus where
 
     {-# INLINABLE verify #-}
     verify :: SetupVerify PlonkPlutus -> Input PlonkPlutus -> Proof PlonkPlutus -> Bool
-    verify SetupBytes{..} InputBytes{..} ProofBytes{..} = bls12_381_finalVerify p1 p2
-        where
-            -- uncompress Setup G1 elements
+    verify SetupBytes{..} InputBytes{..} ProofBytes{..} =
+        let -- uncompress Setup G1 elements
             g0    = bls12_381_G1_uncompress g0'
             h0    = bls12_381_G2_uncompress h0'
             h1    = bls12_381_G2_uncompress h1'
@@ -91,7 +92,7 @@ instance NonInteractiveProof PlonkPlutus where
             u = F . byteStringToInteger LittleEndian . blake2b_224 $ t4 <> proof1' <> proof2'
 
             -- common varibles for r0, d, f, e
-            
+
             xi_n = xi `powTwo` pow
             xi_m_one = xi_n - one
 
@@ -113,7 +114,7 @@ instance NonInteractiveProof PlonkPlutus where
 
             -- final calculations
             r0 =
-                  pubInput * lagrange1_xi
+                  negate pubInput * lagrange1_xi
                 - alphaSquare * lagrange1_xi
                 - alphaEvalZOmega
                     * gamma_beta_a_s1
@@ -157,7 +158,9 @@ instance NonInteractiveProof PlonkPlutus where
                 + v * (s1_xi
                 + v * s2_xi))))
                 + u * z_xi
-                ) `mul` g0
+                ) `mul` g0 -- bls12_381_G1_generator ?
 
-            p1 = bls12_381_millerLoop (xi `mul` proof1 + (u * xi * omega) `mul` proof2 + f - e) h0
+            p1 = bls12_381_millerLoop (xi `mul` proof1 + (u * xi * omega) `mul` proof2 + f - e) h0 -- bls12_381_G2_generator ?
             p2 = bls12_381_millerLoop (proof1 + u `mul` proof2) h1
+
+        in bls12_381_finalVerify p1 p2 && lagsInv * F n * (xi - omega) == one
