+++ zkfold-cardano.cabal
@@ -67,11 +67,16 @@ library
   default-language: Haskell2010
   exposed-modules:
     ZkFold.Cardano.Examples.EqualityCheck
-    ZkFold.Cardano.ForwardingScripts
+    ZkFold.Cardano.Scripts.ForwardingScripts
+    ZkFold.Cardano.Scripts.PlonkVerifier
+    ZkFold.Cardano.Scripts.SymbolicVerifier
     ZkFold.Cardano.Plonk
-    ZkFold.Cardano.PlonkVerifier
-    ZkFold.Cardano.SymbolicVerifier
     ZkFold.Cardano.Plonk.OnChain
+    ZkFold.Cardano.Plonk.OnChain.BLS12_381.F
+    ZkFold.Cardano.Plonk.OnChain.BLS12_381.G1
+    ZkFold.Cardano.Plonk.OnChain.BLS12_381.G2
+    ZkFold.Cardano.Plonk.OnChain.Data
+    ZkFold.Cardano.Plonk.OnChain.Utils
     ZkFold.Cardano.Plonk.OffChain
   build-depends:
     aeson                          ,
